+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - AddReturnPutItemOutput() to accept multiple PutItemOutputs
 - AddReturnQueryOutput() to accept multiple QueryOutputs
 - QueryPages() implemented
+- QueryPagesWithContext() implemented
 
 ## [0.0.2] - 2021-03-07
 ### Added